+++ work-webapi/src/main/java/cn/bjtc/view/SyslogView.java
@@ -5,7 +5,7 @@ public class SyslogView extends Pagination{
 	private Integer logtype;
 	private Integer operator;// 
 	private String content;//操作人
-	private Integer createtime;// 
+	private Long createtime;// 
 	
 	public Integer getId() {
 		return id;
@@ -31,10 +31,10 @@ public class SyslogView extends Pagination{
 	public void setContent(String content) {
 		this.content = content;
 	}
-	public Integer getCreatetime() {
+	public Long getCreatetime() {
 		return createtime;
 	}
-	public void setCreatetime(Integer createtime) {
+	public void setCreatetime(Long createtime) {
 		this.createtime = createtime;
 	}
